new ram with tests, no delay implemented yet
